fix(indexer): support non-Linux resource planning
Provide a conservative GOMAXPROCS-based host capacity fallback for desktop builds and normalize Unix stat identity widths across Linux and Darwin.

+++ internal/libraryindex/fileid_unix.go
@@ -9,7 +9,10 @@ import (
 
 func fileIdentity(info fs.FileInfo) (uint64, uint64) {
 	if stat, ok := info.Sys().(*syscall.Stat_t); ok {
-		return stat.Dev, stat.Ino
+		// Stat_t field widths vary across Unix targets (for example, Dev is
+		// int32 on Darwin and uint64 on Linux). Normalize the persisted identity
+		// explicitly so every supported build compiles with the same schema.
+		return uint64(stat.Dev), uint64(stat.Ino)
 	}
 	return 0, 0
 }
